Tidied some pieces up
