+++ src/constants.js
@@ -1,0 +1,212 @@
+//Object of keys
+const mappings = {
+    8: 'backspace',
+    9: 'tab',
+    12: 'clear',
+    13: 'enter',
+    16: 'shift',
+    17: 'ctrl',
+    18: 'alt',
+    19: 'pause break',
+    20: 'caps lock',
+    27: 'escape',
+    32: 'space',
+    33: 'page up',
+    34: 'page down',
+    35: 'end',
+    36: 'home',
+    37: 'left arrow',
+    38: 'up arrow',
+    39: 'right arrow',
+    40: 'down arrow',
+    41: 'select',
+    42: 'print',
+    43: 'execute',
+    44: 'print screen',
+    45: 'insert',
+    46: 'delete',
+    47: 'help',
+    48: '0',
+    49: '1',
+    50: '2',
+    51: '3',
+    52: '4',
+    53: '5',
+    54: '6',
+    55: '7',
+    56: '8',
+    57: '9',
+    58: ':',
+    59: 'semicolon',
+    60: '<',
+    61: 'equals',
+    64: '@',
+    65: 'a',
+    66: 'b',
+    67: 'c',
+    68: 'd',
+    69: 'e',
+    70: 'f',
+    71: 'g',
+    72: 'h',
+    73: 'i',
+    74: 'j',
+    75: 'k',
+    76: 'l',
+    77: 'm',
+    78: 'n',
+    79: 'o',
+    80: 'p',
+    81: 'q',
+    82: 'r',
+    83: 's',
+    84: 't',
+    85: 'u',
+    86: 'v',
+    87: 'w',
+    88: 'x',
+    89: 'y',
+    90: 'z',
+    91: 'windows key left',
+    92: 'windows key right',
+    93: 'menu right',
+    95: 'sleep',
+    96: 'numpad 0',
+    97: 'numpad 1',
+    98: 'numpad 2',
+    99: 'numpad 3',
+    100: 'numpad 4',
+    101: 'numpad 5',
+    102: 'numpad 6',
+    103: 'numpad 7',
+    104: 'numpad 8',
+    105: 'numpad 9',
+    106: 'multiply',
+    107: 'add',
+    108: 'numpad period',
+    109: 'subtract',
+    110: 'decimal point',
+    111: 'divide',
+    112: 'f1',
+    113: 'f2',
+    114: 'f3',
+    115: 'f4',
+    116: 'f5',
+    117: 'f6',
+    118: 'f7',
+    119: 'f8',
+    120: 'f9',
+    121: 'f10',
+    122: 'f11',
+    123: 'f12',
+    124: 'f13',
+    125: 'f14',
+    126: 'f15',
+    127: 'f16',
+    128: 'f17',
+    129: 'f18',
+    130: 'f19',
+    131: 'f20',
+    132: 'f21',
+    133: 'f22',
+    134: 'f23',
+    135: 'f24',
+    136: 'f25',
+    137: 'f26',
+    138: 'f27',
+    139: 'f28',
+    140: 'f29',
+    141: 'f30',
+    142: 'f31',
+    143: 'f32',
+    144: 'num lock',
+    145: 'scroll lock',
+    160: '^',
+    161: '!',
+    162: '؛',
+    163: '#',
+    164: '$',
+    166: 'page backward',
+    167: 'page forward',
+    168: 'refresh',
+    170: '*',
+    171: '~ + * key',
+    172: 'home key',
+    173: 'minus',
+    174: 'decrease volume level',
+    175: 'increase volume level',
+    176: 'next',
+    177: 'previous',
+    178: 'stop',
+    179: 'play',
+    181: 'mute',
+    182: 'decrease volume level',
+    183: 'increase volume level',
+    186: 'semi-colon / ñ',
+    187: 'equal sign',
+    188: 'comma',
+    189: 'dash',
+    190: 'period',
+    191: 'forward slash',
+    192: 'tilde',
+    194: 'numpad period',
+    219: 'open bracket',
+    220: 'back slash',
+    221: 'close bracket',
+    222: 'single quote',
+    223: '`'
+};
+
+//Object of button codes
+const buttons = {
+    0 : "dpad up",
+    1 : "dpad down",
+    2 : "dpad left",
+    3 : "dpad right",
+    4 : "start",
+    5 : "back",
+    6 : "left thumbstick",
+    7 : "right thumbstick",
+    8 : "left bumper",
+    9 : "right bumper",
+    10 : "a",
+    11 : "b",
+    12 : "x",
+    13 : "y"
+}
+
+const skullEnums = {
+    "gruntBirthday" : 1,
+    "unk" : 2,
+    "unk" : 3,
+    "unk" : 4,
+    "unk" : 5,
+    "blind" : 6,
+    "unk" : 7,
+    "blackeye" : 8,
+    "unk" : 9,
+    "sputnik" : 10,
+    "unk" : 11,
+    "unk" : 12,
+    "unk" : 13,
+    "unk" : 14,
+    "bandana" : 15,
+    "boom" : 16,
+    "eyePatch" : 17,
+    "unk" : 18,
+    "foreign" : 19,
+    "gruntFuneral" : 20,
+    "unk" : 21,
+    "recession" : 22,
+    "malfunction" : 23,
+    "streaking" : 24,
+    "unk" : 25,
+    "scarab" : 26,
+    "feather" : 28
+}
+
+module.exports = {
+    mappings,
+    buttons,
+    skullEnums,
+}
